StreamlitApp/requirements.txt
streamlit
pillow
numpy
pandas
matplotlib
scikit-learn
scikit-image
